Made council positions display on council directory page
Added code that will rewrite the output for the council roles
in the council directory view.

Also added the code that was in the style.css file on the
playscape website, however I have no idea what this code does
so if someone could comment it I would greatly appreciate it.