Added haddrole module, which automagically gives users the honors roles if they are an honors student.
WARNING: This module does not currently check if the user is a honors student.
DO NOT use this module on any website except a local one.
This module will give all users the role specified in the code. This brings me to my next point:
At the current state of the module, you must manually set the rid in the haddrole.module file to the correct rid. Until you do this nothing will happen.