Test and fix Comp Learning form.
Found bugs in db.inc and forms.inc and fixed them.  Also a couple
changes to the database again.  As before, check your database.

The code will still allow you to submit multiple complearning
submissions.  This is due to an unwritten feature in
hstatus_load_user.  For now, this will be ignored b/c it is good
for testing.