Tested and fixed bugs with new cl system
The only known outstanding issue is the fact that the system will
allow you to make submissions even if you have unreviewed
submissions in the system.  This is due to unimplemented features
in hstatus_load_user and is probably desirable from a development
standpoint.

The system is probably also vulnerable to XSS injection attacks.